requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
click==8.1.8
deprecation==2.1.0
fastapi==0.115.7
ffmpeg==1.4
frozenlist==1.5.0
gotrue==2.11.2
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.1.0
packaging==24.2
postgrest==0.19.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
realtime==2.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
storage3==0.11.2
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==13.1
yarl==1.18.3
huggingface_hub==0.21.4
